Move the WSGI application to a separate directory
To quote the mod_wsgi docs,
https://code.google.com/p/modwsgi/wiki/IntegrationWithDjango ,
``Note that you should not go placing the 'django.wsgi' file in the
same directory as the 'settings.py' file, always use a subdirectory.
This is because Apache is being configured to allow serving of files
from that subdirectory. If you put it in the same directory as the
'settings.py' file, you would be inadvertantly allowing Apache to serve
up your source code if someone accidentally mapped a URL to some parent
directory of your project.''