Move WSGI application to separate directory
(This is the corrected version of commit ab17403).

To quote the mod_wsgi docs,
https://code.google.com/p/modwsgi/wiki/IntegrationWithDjango ,
``Note that you should not go placing the 'django.wsgi' file in the
same directory as the 'settings.py' file, always use a subdirectory.
This is because Apache is being configured to allow serving of files
from that subdirectory. If you put it in the same directory as the
'settings.py' file, you would be inadvertantly allowing Apache to serve
up your source code if someone accidentally mapped a URL to some parent
directory of your project.''

`website/settings.py` has been changed to account for the updated
file path to `wsgi.py` too.

Furthermore, `wsgi/__init__.py` was added so that Django can properly
import the WSGI application from its new directory when calling
calling `import clearinghouse.wsgi.wsgi.application`